src/backend/requirements.txt
django
pillow
tablib
openpyxl
pandas
pypdf2
reportlab
python-docx
cryptography
django-rest-knox
django-import-export
django-hitcount
axios
djangorestframework
docx